+++ src/main/java/net/geforcemods/securitycraft/SecurityCraftClient.java
@@ -1,5 +1,6 @@
 package net.geforcemods.securitycraft;
 
+import net.geforcemods.securitycraft.compat.ium.IumCompat;
 import net.neoforged.api.distmarker.Dist;
 import net.neoforged.fml.ModContainer;
 import net.neoforged.fml.common.Mod;
@@ -9,8 +10,11 @@ import net.neoforged.neoforge.client.gui.IConfigScreenFactory;
 
 @Mod(value = SecurityCraft.MODID, dist = Dist.CLIENT)
 public class SecurityCraftClient {
+	public static final IumCompat INSTALLED_IUM_MOD = IumCompat.getInstalledIumMod();
+
 	public SecurityCraftClient(ModContainer container) {
 		container.registerExtensionPoint(IConfigScreenFactory.class, ConfigurationScreen::new);
 		container.registerConfig(ModConfig.Type.CLIENT, ConfigHandler.CLIENT_SPEC);
+		INSTALLED_IUM_MOD.resolve();
 	}
 }