+++ src/main/java/net/geforcemods/securitycraft/mixin/camera/ClientChunkCacheMixin.java
@@ -10,8 +10,7 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
-import net.geforcemods.securitycraft.SecurityCraft;
-import net.geforcemods.securitycraft.compat.embeddium.EmbeddiumCompat;
+import net.geforcemods.securitycraft.SecurityCraftClient;
 import net.geforcemods.securitycraft.entity.camera.CameraController;
 import net.geforcemods.securitycraft.misc.IChunkStorageProvider;
 import net.geforcemods.securitycraft.util.PlayerUtils;
@@ -94,9 +93,7 @@ public abstract class ClientChunkCacheMixin implements IChunkStorageProvider {
 			if (chunk != null && chunk.getPos().x == x && chunk.getPos().z == z) {
 				NeoForge.EVENT_BUS.post(new ChunkEvent.Unload(chunk));
 				cameraStorage.replace(i, chunk, null);
-
-				if (SecurityCraft.IS_EMBEDDIUM_INSTALLED)
-					EmbeddiumCompat.onChunkStatusRemoved(level, x, z);
+				SecurityCraftClient.INSTALLED_IUM_MOD.onChunkStatusRemoved(level, x, z);
 			}
 		}
 	}
@@ -123,10 +120,7 @@ public abstract class ClientChunkCacheMixin implements IChunkStorageProvider {
 				chunk.replaceWithPacketData(buffer, chunkTag, tagOutputConsumer);
 
 			level.onChunkLoaded(chunkPos);
-
-			if (SecurityCraft.IS_EMBEDDIUM_INSTALLED)
-				EmbeddiumCompat.onChunkStatusAdded(level, x, z);
-
+			SecurityCraftClient.INSTALLED_IUM_MOD.onChunkStatusAdded(level, x, z);
 			NeoForge.EVENT_BUS.post(new ChunkEvent.Load(chunk, false));
 			callback.setReturnValue(chunk);
 		}