+++ Changelog.md
@@ -40,6 +40,7 @@
 - Fix: Potential crash
 - Fix: Setting the signal length option to 0 on a Keypad Trapdoor makes the trapdoor not open at all
 - Fix: Block Pocket and Block Change Detector outline/block highlights are drawn too thin at certain angles
+- Fix: Cameras do not work from far away when using Sodium
 
 --------------------------Changelog for v1.9.10-beta9 of SecurityCraft--------------------------